refactor: 알아보기쉽게 keydown을 분리하였습니다.
