+++ Box42/Main/keyDown+BoxViewController.swift
@@ -1,0 +1,33 @@
+//
+//  keyDown+BoxViewController.swift
+//  Box42
+//
+//  Created by Chanhee Kim on 8/21/23.
+//
+
+import AppKit
+
+extension BoxViewController {
+    override func keyDown(with event: NSEvent) {
+        print(event.keyCode)
+        if event.keyCode == 1 {
+            StorageConfig.shared.setThreshold(.percentage50)
+            StorageConfig.shared.setPeriod(.period10s)
+        }
+        if event.keyCode == 2 {
+            //            SdtorageConfig.shared.setThreshold(.percentage30)
+            DispatchQueue.main.async {
+                StorageConfig.shared.setThreshold(.percentage30)
+            }
+            StorageConfig.shared.setPeriod(.period1s)
+        }
+        
+        
+        if event.keyCode == 53 { // Escape 키의 keyCode는 53입니다.
+            print("escape")
+            menubarVCDelegate?.toggleWindow(sender: nil)
+        } else {
+            super.keyDown(with: event) // 기타 키를 처리하기 위해 상위 클래스에게 전달
+        }
+    }
+}
